Validations refactor
