+++ lib/grape/validations/validators/values.rb
@@ -1,9 +1,8 @@
 module Grape
   module Validations
-    class ValuesValidator < Validator
+    class ValuesValidator < Base
       def initialize(attrs, options, required, scope)
         @values = options
-        @required = required
         super
       end
