+++ spec/grape/validations/attributes_iterator_spec.rb
@@ -1,0 +1,4 @@
+require 'spec_helper'
+
+describe Grape::Validations::AttributesIterator do
+end
