+++ spec/grape/validations/params_scope_spec.rb
@@ -1,0 +1,4 @@
+require 'spec_helper'
+
+describe Grape::Validations::ParamsScope do
+end
