+++ spec/grape/validations/validators/at_least_one_of_spec.rb
