+++ spec/grape/validations/validators/exactly_one_of_spec.rb
